+++ CHANGELOG.md
@@ -4,9 +4,12 @@ All notable changes to this project will be documented in this file.
 **ATTN**: This project uses [semantic versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [v0.10.0] - 2022-07-08
 ### Added
 - Added player hit record to pvp log file.
 - Added change safehouse owner action to safehouse log file.
+- Added logs for send safehouse invite action.
 
 ## [v0.9.0] - 2022-06-09
 ### Added
@@ -108,8 +111,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Add basic implementation.
 
-[Unreleased]: https://github.com/openzomboid/log-extender/compare/v0.9.0...HEAD
-[v0.8.0]: https://github.com/openzomboid/log-extender/compare/v0.8.0...v0.9.0
+[Unreleased]: https://github.com/openzomboid/log-extender/compare/v0.10.0...HEAD
+[v0.10.0]: https://github.com/openzomboid/log-extender/compare/v0.9.0...v0.10.0
+[v0.9.0]: https://github.com/openzomboid/log-extender/compare/v0.8.0...v0.9.0
 [v0.8.0]: https://github.com/openzomboid/log-extender/compare/v0.7.1...v0.8.0
 [v0.7.1]: https://github.com/openzomboid/log-extender/compare/v0.7.0...v0.7.1
 [v0.7.0]: https://github.com/openzomboid/log-extender/compare/v0.6.0...v0.7.0